Moved Body from component to area
Body was a complete section of the app so it was rather unfitting to
keep it in components as it is not an individual component. It was
therefore moved to Areas.
